Completed Assignment

+++ app/controllers/toys_controller.rb
@@ -6,20 +6,30 @@ class ToysController < ApplicationController
     render json: toys
   end
 
+  def show
+    toy = Toy.find_by(id: params[:id])
+    render json: toy
+  end
+
   def create
-    toy = Toys.create(toy_params)
+    toy = Toy.create(toy_params)
     render json: toy, status: :created
   end
 
   def update
     toy = Toy.find_by(id: params[:id])
     toy.update(toy_params)
+    render json: toy, status: :accepted
   end
 
   def destroy
     toy = Toy.find_by(id: params[:id])
-    toy.destroy
-    head :no_content
+    if toy
+      toy.destroy
+      render json: {}
+    else 
+      render json: {error:"Toy not found"}, status: :not_found
+    end
   end
 
   private
